+++ modules/self-hosted/default.nix
@@ -31,6 +31,7 @@
     # Apps and utilities
     ./cloakbrowser.nix
     ./hermes.nix
+    ./honcho.nix
     ./bentopdf.nix
     ./convertx.nix
     ./it-tools.nix