+++ modules/self-hosted/secrets.nix
@@ -84,4 +84,14 @@
     mode = "0400";
   };
 
+  sops.secrets."honcho-secrets" = {
+    sopsFile = ../../secrets.yaml;
+    mode = "0400";
+  };
+
+  sops.secrets."litellm-secrets" = {
+    sopsFile = ../../secrets.yaml;
+    mode = "0400";
+  };
+
 }